--- e2e/test/scenarios/admin/databases/default-sample-database.cy.spec.js
@@ -1,298 +1,0 @@
-import { SAMPLE_DB_ID } from "e2e/support/cypress_data";
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import { ORDERS_QUESTION_ID } from "e2e/support/cypress_sample_instance_data";
-import { modal, popover, restore } from "e2e/support/helpers";
-import {
-  createMetric,
-  createSegment,
-} from "e2e/support/helpers/e2e-table-metadata-helpers";
-
-import { visitDatabase } from "./helpers/e2e-database-helpers";
-
-const { ORDERS_ID, ORDERS } = SAMPLE_DATABASE;
-
-describe("scenarios > admin > databases > sample database", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-    cy.intercept("PUT", "/api/database/*").as("databaseUpdate");
-  });
-
-  it("database settings", () => {
-    visitDatabase(SAMPLE_DB_ID);
-    // should not display a setup help card
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Need help connecting?").should("not.exist");
-
-    cy.log(
-      "should not be possible to change database type for the Sample Database (metabase#16382)",
-    );
-    cy.findByLabelText("Database type")
-      .should("have.text", "H2")
-      .and("be.disabled");
-
-    cy.log("should correctly display connection settings");
-    cy.findByLabelText("Display name").should("have.value", "Sample Database");
-    cy.findByLabelText("Connection String")
-      .should("have.attr", "value")
-      .and("contain", "sample-database.db");
-
-    cy.log("should be possible to modify the connection settings");
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Show advanced options").click();
-    // `auto_run_queries` toggle should be ON by default
-    cy.findByLabelText("Rerun queries for simple explorations")
-      .should("have.attr", "aria-checked", "true")
-      .click();
-    // Reported failing in v0.36.4
-    cy.log(
-      "should respect the settings for automatic query running (metabase#13187)",
-    );
-    cy.findByLabelText("Rerun queries for simple explorations").should(
-      "have.attr",
-      "aria-checked",
-      "false",
-    );
-
-    cy.log("change the metadata_sync period");
-    cy.findByLabelText("Choose when syncs and scans happen").click();
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Hourly").click();
-    popover().within(() => {
-      cy.findByText("Daily").click({ force: true });
-    });
-
-    // "lets you change the cache_field_values period"
-    cy.findByLabelText("Never, I'll do this manually if I need to").should(
-      "have.attr",
-      "aria-selected",
-      "true",
-    );
-
-    cy.findByLabelText("Regularly, on a schedule")
-      .click()
-      .within(() => {
-        cy.findByText("Daily").click();
-      });
-    popover().findByText("Weekly").click();
-
-    cy.button("Save changes").click();
-    cy.wait("@databaseUpdate").then(({ response: { body } }) => {
-      expect(body.details["let-user-control-scheduling"]).to.equal(true);
-      expect(body.schedules.metadata_sync.schedule_type).to.equal("daily");
-      expect(body.schedules.cache_field_values.schedule_type).to.equal(
-        "weekly",
-      );
-    });
-    cy.button("Success");
-
-    // "lets you change the cache_field_values to 'Only when adding a new filter widget'"
-    cy.findByLabelText("Only when adding a new filter widget").click();
-    cy.button("Save changes", { timeout: 10000 }).click();
-    cy.wait("@databaseUpdate").then(({ response: { body } }) => {
-      expect(body.is_full_sync).to.equal(false);
-      expect(body.is_on_demand).to.equal(true);
-    });
-
-    // and back to never
-    cy.findByLabelText("Never, I'll do this manually if I need to").click();
-    cy.button("Save changes", { timeout: 10000 }).click();
-    cy.wait("@databaseUpdate").then(({ response: { body } }) => {
-      expect(body.is_full_sync).to.equal(false);
-      expect(body.is_on_demand).to.equal(false);
-    });
-  });
-
-  it("database actions sidebar", () => {
-    cy.intercept("POST", `/api/database/${SAMPLE_DB_ID}/sync_schema`).as(
-      "sync_schema",
-    );
-    cy.intercept("POST", `/api/database/${SAMPLE_DB_ID}/rescan_values`).as(
-      "rescan_values",
-    );
-    cy.intercept("POST", `/api/database/${SAMPLE_DB_ID}/discard_values`).as(
-      "discard_values",
-    );
-    cy.intercept("GET", `/api/database/${SAMPLE_DB_ID}/usage_info`).as(
-      "usage_info",
-    );
-    cy.intercept("DELETE", `/api/database/${SAMPLE_DB_ID}`).as("delete");
-    // model
-    cy.request("PUT", `/api/card/${ORDERS_QUESTION_ID}`, { type: "model" });
-    // Create a segment through API
-    createSegment({
-      name: "Small orders",
-      description: "All orders with a total under $100.",
-      table_id: ORDERS_ID,
-      definition: {
-        "source-table": ORDERS_ID,
-        aggregation: [["count"]],
-        filter: ["<", ["field", ORDERS.TOTAL, null], 100],
-      },
-    });
-
-    // metric
-    createMetric({
-      name: "Revenue",
-      description: "Sum of orders subtotal",
-      table_id: ORDERS_ID,
-      definition: {
-        "source-table": ORDERS_ID,
-        aggregation: [["sum", ["field", ORDERS.SUBTOTAL, null]]],
-      },
-    });
-
-    visitDatabase(SAMPLE_DB_ID);
-
-    // lets you trigger the manual database schema sync
-    cy.button("Sync database schema now").click();
-    cy.wait("@sync_schema");
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Sync triggered!");
-
-    // lets you trigger the manual rescan of field values
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Re-scan field values now").click();
-    cy.wait("@rescan_values");
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Scan triggered!");
-
-    // lets you discard saved field values
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Danger Zone")
-      .parent()
-      .as("danger")
-      .within(() => {
-        cy.button("Discard saved field values").click();
-      });
-    modal().within(() => {
-      cy.findByRole("heading").should(
-        "have.text",
-        "Discard saved field values",
-      );
-      cy.findByText("Are you sure you want to do this?");
-      cy.button("Yes").click();
-    });
-    cy.wait("@discard_values");
-
-    // lets you remove the Sample Database
-    cy.get("@danger").within(() => {
-      cy.button("Remove this database").click();
-      cy.wait("@usage_info");
-    });
-
-    modal().within(() => {
-      cy.button("Delete this content and the DB connection")
-        .as("deleteButton")
-        .should("be.disabled");
-      cy.findByLabelText(/Delete [0-9]* saved questions?/)
-        .should("not.be.checked")
-        .click()
-        .should("be.checked");
-      cy.findByLabelText(/Delete [0-9]* models?/)
-        .should("not.be.checked")
-        .click()
-        .should("be.checked");
-      cy.findByLabelText(/Delete [0-9]* metrics?/)
-        .should("not.be.checked")
-        .click()
-        .should("be.checked");
-      cy.findByLabelText(/Delete [0-9]* segments?/)
-        .should("not.be.checked")
-        .click()
-        .should("be.checked");
-      cy.findByText(
-        "This will delete every saved question, model, metric, and segment you’ve made that uses this data, and can’t be undone!",
-      );
-
-      cy.get("@deleteButton").should("be.disabled");
-
-      cy.findByPlaceholderText("Are you completely sure?")
-        .type("Sample Database")
-        .blur();
-
-      cy.intercept("GET", "/api/database").as("fetchDatabases");
-      cy.get("@deleteButton").should("be.enabled").click();
-      cy.wait(["@delete", "@fetchDatabases"]);
-    });
-
-    cy.location("pathname").should("eq", "/admin/databases/"); // FIXME why the trailing slash?
-    cy.intercept("POST", "/api/database/sample_database").as("sample_database");
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.contains("Bring the sample database back", {
-      timeout: 10000,
-    }).click();
-    cy.wait("@sample_database");
-
-    cy.findAllByRole("cell").contains("Sample Database").click();
-    const newSampleDatabaseId = SAMPLE_DB_ID + 1;
-    cy.location("pathname").should(
-      "eq",
-      `/admin/databases/${newSampleDatabaseId}`,
-    );
-  });
-
-  it("updates databases list in Browse databases after bringing sample database back", () => {
-    cy.intercept("GET", "/api/database").as("loadDatabases");
-    cy.intercept("POST", "/api/database/sample_database").as(
-      "restoreSampleDatabase",
-    );
-    cy.intercept("GET", "/api/database/*/usage_info").as(
-      "loadDatabaseUsageInfo",
-    );
-    cy.intercept("GET", "/api/database/*").as("loadDatabase");
-    cy.intercept("DELETE", "/api/database/*").as("deleteDatabase");
-
-    cy.visit("/admin/databases");
-    cy.wait("@loadDatabases");
-    cy.findByTestId("database-list").within(() => {
-      cy.findByText("Sample Database").click();
-    });
-    cy.wait("@loadDatabase");
-
-    cy.button("Remove this database").click();
-    cy.wait("@loadDatabaseUsageInfo");
-    modal().within(() => {
-      cy.findByLabelText(/Delete \d+ saved questions?/).click();
-      cy.findByLabelText(/Delete \d+ model?/).click();
-      cy.findByTestId("database-name-confirmation-input").type(
-        "Sample Database",
-      );
-      cy.findByText("Delete this content and the DB connection").click();
-      cy.wait("@deleteDatabase");
-    });
-
-    cy.findByTestId("database-list").within(() => {
-      cy.findByText("Sample Database").should("not.exist");
-    });
-
-    cy.findByTestId("exit-admin").click();
-
-    cy.wait("@loadDatabases");
-    cy.findByTestId("main-navbar-root").within(() => {
-      cy.findByLabelText("Browse databases").should("not.exist");
-    });
-
-    cy.visit("/admin/databases");
-    cy.findByTestId("database-list").within(() => {
-      cy.findByText("Bring the sample database back").click();
-      cy.wait("@restoreSampleDatabase");
-    });
-
-    cy.findByTestId("database-list").within(() => {
-      cy.findByText("Sample Database").should("exist");
-    });
-
-    cy.findByTestId("exit-admin").click();
-
-    cy.wait("@loadDatabases");
-    cy.findByTestId("main-navbar-root").within(() => {
-      cy.findByLabelText("Browse databases").should("exist");
-      cy.findByLabelText("Browse databases").click();
-    });
-
-    cy.findByTestId("database-browser").within(() => {
-      cy.findByText("Sample Database").should("exist");
-    });
-  });
-});
