--- e2e/test/scenarios/admin/databases/reproductions/33035-group-translations.cy.spec.js
@@ -1,17 +1,0 @@
-import { SAMPLE_DB_ID } from "e2e/support/cypress_data";
-import { restore } from "e2e/support/helpers";
-
-describe("issue 33035", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-    cy.request("GET", "/api/user/current").then(({ body: { id: user_id } }) => {
-      cy.request("PUT", `/api/user/${user_id}`, { locale: "de" });
-    });
-  });
-
-  it("databases page should work in a non-default locale (metabase#33035)", () => {
-    cy.visit(`/admin/permissions/data/database/${SAMPLE_DB_ID}`);
-    cy.findByRole("main").findByText("Orders").should("be.visible");
-  });
-});
