+++ e2e/test/scenarios/admin/subscription.cy.spec.js
