--- e2e/test/scenarios/admin/troubleshooting/help.cy.spec.js
@@ -1,72 +1,0 @@
-import {
-  onlyOnOSS,
-  describeEE,
-  restore,
-  setupMetabaseCloud,
-  setTokenFeatures,
-} from "e2e/support/helpers";
-
-describe("scenarios > admin > troubleshooting > help", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-  });
-
-  // Unskip when mocking Cloud in Cypress is fixed (#18289)
-  it.skip("should add the support link when running Metabase Cloud", () => {
-    setupMetabaseCloud();
-    cy.visit("/admin/troubleshooting/help");
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Metabase Admin");
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Contact support");
-  });
-});
-
-describe("scenarios > admin > troubleshooting > help", { tags: "@OSS" }, () => {
-  beforeEach(() => {
-    onlyOnOSS();
-
-    restore();
-    cy.signInAsAdmin();
-  });
-
-  it("should link `Get Help` to help", () => {
-    cy.visit("/admin/troubleshooting/help");
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Metabase Admin");
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Get Help")
-      .parents("a")
-      .should("have.prop", "href")
-      .and(
-        "match",
-        /^https:\/\/www\.metabase\.com\/help\?utm_source=in-product&utm_medium=troubleshooting&utm_campaign=help&instance_version=v(?:(?!diag=).)+$/,
-      );
-  });
-});
-
-describeEE("scenarios > admin > troubleshooting > help (EE)", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-    setTokenFeatures("all");
-  });
-
-  it("should link `Get Help` to help-premium", () => {
-    cy.visit("/admin/troubleshooting/help");
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Metabase Admin");
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Get Help")
-      .parents("a")
-      .should("have.prop", "href")
-      .and(
-        "match",
-        /^https:\/\/www\.metabase\.com\/help-premium\?utm_source=in-product&utm_medium=troubleshooting&utm_campaign=help&instance_version=v.+&diag=%7B.+%7D$/,
-      );
-  });
-});
